+++ desktop/cmp/input/Select.ts
@@ -434,11 +434,12 @@ class SelectInputModel extends HoistInputModel {
     // (Exception for a null value, which we will only accept if explicitly present in options.)
     override toInternal(external) {
         if (this.multiMode) {
-            if (external == this.emptyValue) external = []; // avoid [null]
-            return castArray(external).map(it => this.findOption(it, !isNil(it)));
+            if (external === null) external = []; // avoid [null]
+            return castArray(external).map(it =>
+                this.findOption(it, !isNil(it) && it?.toString() !== this.emptyValue?.toString())
+            );
         }
-
-        return this.findOption(external, external !== this.emptyValue);
+        return this.findOption(external, !isNil(external) && external !== this.emptyValue);
     }
 
     private findOption(value, createIfNotFound, options = this.internalOptions) {